Fix; Duplicate Navigation Implementation
