+++ app/(tabs)/search.tsx
@@ -1,0 +1,2 @@
+import SearchScreen from '../../src/screens/SearchScreen';
+export default SearchScreen;
