+++ app/settings.tsx
@@ -1,0 +1,2 @@
+import SettingsScreen from '../src/screens/SettingsScreen';
+export default SettingsScreen;
